cli/pyproject.toml
[project]
name = "context-os-events"
version = "0.1.0"
description = "Event capture layer for Context OS - tracks file events, Claude sessions, and git commits"
requires-python = ">=3.10"

dependencies = [
    "watchdog>=3.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "schedule>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
context-os = "context_os_events.cli:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
